arfx-collect can: load channel names from files, chunk large entries

+++ arfx/collect.py
@@ -90,6 +90,23 @@ def check_entry_consistency(arfp, entries=None, channels=None, predicate=any_typ
     return entry_names, channel_props
 
 
+def iter_entry_chunks(entry, channels, predicate, size):
+    """ Iterate through the datasets in entry (that match predicate), yielding chunks """
+    props = channel_properties(entry, channels, predicate)
+    nchannels = len(props)
+    nsamples = first(props, operator.itemgetter("samples"))
+    dtype = first(props, operator.itemgetter("dtype"))
+    log.info("    - '%s' -> %d samples", entry.name, nsamples)
+    for i in range(0, nsamples, size):
+        n = min(size, nsamples - i)
+        out = np.empty((n, nchannels), dtype=dtype)
+        log.debug("            - %d - %d", i, i + n)
+        for j, chan_name in enumerate(props):
+            dset = entry[chan_name]
+            out[:, j] = dset[i:i+n]
+        yield out
+
+
 def collect_sampled_script(argv=None):
     from natsort import natsorted
     import argparse
@@ -102,11 +119,18 @@ def collect_sampled_script(argv=None):
     p.add_argument('--version', action="version",
                    version="%(prog)s " + __version__)
     p.add_argument('-v', '--verbose', help="show verbose log messages", action="store_true")
+    p.add_argument("--dry-run", action="store_true",
+                   help="check entry consistency but don't write file")
 
-    p.add_argument("-d", "--dtype", help="convert data to specified type (default is to use as stored)")
+    p.add_argument("-d", "--dtype",
+                   help="convert data to specified type (default is to use as stored)")
+    p.add_argument("--chunk-size", type=int, default=1 << 22,
+                   help="minimum chunk size for processing long datasets")
     # p.add_argument("-b", "--bark", help="output bark meta.yml file", action="store_true")
-    p.add_argument("-c", "--channels", help="list of channels to unpack (default all)",
-                   metavar='CHANNEL', nargs="+")
+    p.add_argument("-c", "--channels", metavar='CHANNEL', nargs="+", default=[],
+                   help="list of channels to unpack (default all)")
+    p.add_argument("-C", "--channel-file",
+                   help="file with list of channels to unpack, one per line")
     p.add_argument('-e', '--entries', help="list of entries to unpack (default all)",
                    metavar='ENTRY', nargs='+')
 
@@ -115,6 +139,14 @@ def collect_sampled_script(argv=None):
 
     args = p.parse_args(argv)
     setup_log(log, args.verbose)
+
+    if args.channel_file is not None:
+        with open(args.channel_file, "rt") as fp:
+            for line in fp:
+                if line.startswith("#"):
+                    continue
+                else:
+                    args.channels.append(line.strip())
 
     with arf.open_file(args.arffile, "r") as arfp:
         log.info("unpacking '%s'", args.arffile)
@@ -141,12 +173,13 @@ def collect_sampled_script(argv=None):
         log.info(" - sampling rate = %f", sampling_rate)
         log.info(" - dtype = '%s'", dtype)
         log.info(" - entries (%d):", nentries)
+        if args.dry_run:
+            log.info(" - dry run, ending script")
+            return
         with io.open(args.outfile, mode="w",
                      sampling_rate=sampling_rate, dtype=dtype, nchannels=nchannels) as ofp:
             for entry_name in natsorted(entry_names):
                 entry = arfp[entry_name]
-                # nsamples = first(entry, operator.attrgetter("shape"))[0]
-                # would be more efficient to preallocate but this is easy
-                data = np.column_stack(entry[cname][:] for cname in natsorted(channel_props))
-                ofp.write(data)
-                log.info("    - '%s' -> %d samples", entry_name, data.shape[0])
+                for chunk in iter_entry_chunks(entry, args.channels, arf.is_time_series, args.chunk_size):
+
+                    ofp.write(chunk.astype(dtype))
